translater/requirements.txt
streamlit
yt-dlp
faster-whisper
pydub
numpy
torch  # faster-whisper에서 내부적으로 사용
ffmpeg-python  # pydub 사용 시 필요할 수 있음